fastapi_example/pyproject.toml
[project]
name = "fastapi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.1",
    "google-cloud-run>=0.12.0",
    "uvicorn>=0.38.0",
]
